Update grades.py
Managed to get the script to find the username and password input fields by changing how it finds the elements. Instead of using "presense_of_element" which didn't work, I used "find_element". I also used the find_all() and find() methods to find attributes with certain ids/class names and extracted the data in those attributes. The problem now is that it's not extracting the data even with everything required given.